fix: move to /plugins/slack/shared
