+++ plugins/slack/shared/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../server/.babelrc"
+}
